requirements.txt
bcrypt==3.2.0
cryptography==36.0.1
fastapi==0.73.0
Jinja2==3.0.3
passlib==1.7.4
pydantic==1.9.0
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.16.0
sqlalchemy==1.4.36
psycopg2==2.9.3